Section 5, Lecture 112: Learnt What is Pose Mode
Learnt What is and How to use Pose Mode

Pose Mode is select the Bone and transform such as Rotation, the Object Mesh will be Move together

If we Connect the Bone in wrong way.
First to Un Parent
Second, move the Bone to Correct Location.
Third Reparent. Select the Object, Than Select the Bone in Pose Mode, Parent (ctrl with P) --> Bone